Add service autoscaling info to status
Inspired by recent addition of cluster autoscaling to metastatus this
will give service autoscaling info in paasta status.

This refactors some of the autoscaling service lib into smaller
functions so that I can easily implement a function to just get the info
on what the autoscaler would do. It also adds a noop function to our
metric providers because they usually update data in zookeeper which
would be undesirable when we just want to know what the autoscaler would
do if it ran now.